# .agent/tmp/robyn_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OAP LangGraph Runtime",
    "version": "0.1.0",
    "description": "\nRobyn-based LangGraph-compatible runtime for Open Agent Platform.\n\nThis API provides endpoints for managing assistants, threads, runs, and the store.\nIt implements the LangGraph API specification for compatibility with LangGraph clients.\n"
  },
...
  "paths": {
...
    "/store/items": {
      "get": {
        "tags": ["Store"],
        "summary": "Get Store Item",
        "description": "Retrieve an item from the store by namespace and key.",
        "operationId": "get_store_item",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "The namespace (dot-separated, e.g., 'users.profiles')."
          },
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "The key within the namespace."
          }
        ],
        "responses": {
          "200": {
            "description": "The requested item",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Item" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Store"],
        "summary": "Put Store Item",
        "description": "Store an item in the key-value store.",
        "operationId": "put_store_item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StorePutRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Item stored successfully" },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Store"],
        "summary": "Delete Store Item",
        "description": "Delete an item from the store.",
        "operationId": "delete_store_item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StoreDeleteRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Item deleted successfully" },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/store/items/search": {
      "post": {
        "tags": ["Store"],
        "summary": "Search Store Items",
        "description": "Search for items in the store.",
        "operationId": "search_store_items",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StoreSearchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SearchItemsResponse" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/store/namespaces": {
      "post": {
        "tags": ["Store"],
        "summary": "List Namespaces",
        "description": "List namespaces in the store.",
        "operationId": "list_namespaces",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreListNamespacesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of namespaces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message describing what went wrong."
          }
        },
        "title": "ErrorResponse",
        "description": "Standard error response format."
      },
...
      "StorePutRequest": {
        "type": "object",
        "required": ["namespace", "key", "value"],
        "properties": {
          "namespace": {
            "type": "array",
            "items": { "type": "string" },
            "title": "Namespace",
            "description": "The namespace for the item (e.g., ['users', 'profiles'])."
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The key for the item within the namespace."
          },
          "value": {
            "type": "object",
            "title": "Value",
            "description": "The value to store (must be JSON-serializable)."
          }
        },
        "title": "StorePutRequest",
        "description": "Request to store an item."
      },
      "StoreDeleteRequest": {
        "type": "object",
        "required": ["namespace", "key"],
        "properties": {
          "namespace": {
            "type": "array",
            "items": { "type": "string" },
            "title": "Namespace",
            "description": "The namespace for the item."
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The key for the item to delete."
          }
        },
        "title": "StoreDeleteRequest",
        "description": "Request to delete an item."
      },
      "StoreSearchRequest": {
        "type": "object",
        "properties": {
          "namespace_prefix": {
            "type": "array",
            "items": { "type": "string" },
            "title": "Namespace Prefix",
            "description": "Filter by namespace prefix."
          },
          "filter": {
            "type": "object",
            "title": "Filter",
            "description": "Filter by value fields."
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 10,
            "description": "Maximum number of results."
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0,
            "description": "Number of results to skip."
          },
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Full-text search query."
          }
        },
        "title": "StoreSearchRequest",
        "description": "Request to search store items."
      },
      "StoreListNamespacesRequest": {
        "type": "object",
        "properties": {
          "prefix": {
            "type": "array",
            "items": { "type": "string" },
            "title": "Prefix",
            "description": "Filter by namespace prefix."
          },
          "suffix": {
            "type": "array",
            "items": { "type": "string" },
            "title": "Suffix",
            "description": "Filter by namespace suffix."
          },
          "max_depth": {
            "type": "integer",
            "title": "Max Depth",
            "description": "Maximum namespace depth to return."
          },
          "limit": { "type": "integer", "title": "Limit", "default": 100 },
          "offset": { "type": "integer", "title": "Offset", "default": 0 }
        },
        "title": "StoreListNamespacesRequest",
        "description": "Request to list namespaces."
      },
      "Item": {
        "type": "object",
        "required": ["namespace", "key", "value"],
        "properties": {
          "namespace": {
            "type": "array",
            "items": { "type": "string" },
            "description": "The namespace of the item."
          },
          "key": { "type": "string", "description": "The key of the item." },
          "value": { "type": "object", "description": "The stored value." },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the item was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the item was last updated."
          }
        },
        "description": "A stored item."
      },
      "SearchItemsResponse": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Item" }
          }
        }
      },
...
    }
  }
}